rename: create 패키지로 분리했던 클래스를 meeting 패키지로 통합
